+++ src/plots/column/adaptor.ts
@@ -16,8 +16,9 @@ import { conversionTag } from '../../adaptor/conversion-tag';
 import { connectedArea } from '../../adaptor/connected-area';
 import { interval } from '../../adaptor/geometries';
 import { flow, transformLabel, deepAssign, findGeometry, adjustYMetaByZero, pick } from '../../utils';
-import { getDataWhetherPecentage } from '../../utils/transform/percent';
+import { getDataWhetherPecentage, getDeepPercent } from '../../utils/transform/percent';
 import { Datum } from '../../types';
+import { brushInteraction } from '../../adaptor/brush';
 import { ColumnOptions } from './types';
 
 /**
@@ -52,15 +53,50 @@ function defaultOptions(params: Params<ColumnOptions>): Params<ColumnOptions> {
  */
 function geometry(params: Params<ColumnOptions>): Params<ColumnOptions> {
   const { chart, options } = params;
-  const { data, columnStyle, color, columnWidthRatio, isPercent, xField, yField, seriesField, tooltip } = options;
+  const {
+    data,
+    columnStyle,
+    color,
+    columnWidthRatio,
+    isPercent,
+    isGroup,
+    isStack,
+    xField,
+    yField,
+    seriesField,
+    groupField,
+    tooltip,
+  } = options;
 
-  chart.data(getDataWhetherPecentage(data, yField, xField, yField, isPercent));
+  const percentData =
+    isPercent && isGroup && isStack
+      ? getDeepPercent(data, yField, [xField, groupField], yField)
+      : getDataWhetherPecentage(data, yField, xField, yField, isPercent);
+
+  let chartData = [];
+
+  // 存在堆叠,并且存在堆叠seriesField分类，并且不存在分组的时候 进行堆叠
+  if (isStack && seriesField && !isGroup) {
+    percentData.forEach((item) => {
+      const stackedItem = chartData.find((v) => v[xField] === item[xField] && v[seriesField] === item[seriesField]);
+      if (stackedItem) {
+        stackedItem[yField] += item[yField] || 0;
+      } else {
+        chartData.push({ ...item });
+      }
+    });
+  } else {
+    chartData = percentData;
+  }
+
+  chart.data(chartData);
 
   // 百分比堆积图，默认会给一个 % 格式化逻辑, 用户可自定义
   const tooltipOptions = isPercent
     ? {
         formatter: (datum: Datum) => ({
-          name: datum[seriesField] || datum[xField],
+          name:
+            isGroup && isStack ? `${datum[seriesField]} - ${datum[groupField]}` : datum[seriesField] || datum[xField],
           value: (Number(datum[yField]) * 100).toFixed(2) + '%',
         }),
         ...tooltip,
@@ -69,6 +105,7 @@ function geometry(params: Params<ColumnOptions>): Params<ColumnOptions> {
 
   const p = deepAssign({}, params, {
     options: {
+      data: chartData,
       widthRatio: columnWidthRatio,
       tooltip: tooltipOptions,
       interval: {
@@ -79,7 +116,7 @@ function geometry(params: Params<ColumnOptions>): Params<ColumnOptions> {
   });
   interval(p);
 
-  return params;
+  return p;
 }
 
 /**
@@ -259,6 +296,7 @@ export function adaptor(params: Params<ColumnOptions>, isBar = false) {
     slider,
     scrollbar,
     label,
+    brushInteraction,
     interaction,
     animation,
     annotation(),