+++ src/plots/radial-bar/utils.ts
@@ -9,3 +9,19 @@ export function getScaleMax(maxAngle: number, yField: string, data: Data): numbe
   }
   return (maxValue * 360) / formatRadian;
 }
+
+/**
+ * 获取堆叠之后的数据
+ */
+export function getStackedData(data: Data, xField: string, yField: string) {
+  const stackedData: Data = [];
+  data.forEach((item) => {
+    const valueItem = stackedData.find((v) => v[xField] === item[xField]);
+    if (valueItem) {
+      valueItem[yField] += item[yField] || null;
+    } else {
+      stackedData.push({ ...item });
+    }
+  });
+  return stackedData;
+}