+++ src/plots/rose/constant.ts
@@ -9,7 +9,6 @@ export const DEFAULT_OPTIONS = deepAssign({}, Plot.getDefaultOptions(), {
   yAxis: false,
   legend: {
     position: 'right',
-    offsetX: -10,
   },
   sectorStyle: {
     stroke: '#fff',