+++ src/plots/sunburst/index.ts
@@ -1,5 +1,6 @@
 import { Plot } from '../../core/plot';
 import { Adaptor } from '../../core/adaptor';
+import { NODE_ANCESTORS_FIELD } from '../../utils/hierarchy/util';
 import { SunburstOptions } from './types';
 import { adaptor } from './adaptor';
 import { DEFAULT_OPTIONS, SUNBURST_ANCESTOR_FIELD, SUNBURST_PATH_FIELD } from './constant';
@@ -20,6 +21,8 @@ export class Sunburst extends Plot<SunburstOptions> {
   static SUNBURST_ANCESTOR_FIELD = SUNBURST_ANCESTOR_FIELD;
   /** 旭日图 节点的路径 */
   static SUNBURST_PATH_FIELD = SUNBURST_PATH_FIELD;
+  /** 节点的祖先节点 */
+  static NODE_ANCESTORS_FIELD = NODE_ANCESTORS_FIELD;
 
   /** 图表类型 */
   public type: string = 'sunburst';