+++ src/utils/index.ts
@@ -13,3 +13,4 @@ export { renderStatistic, renderGaugeStatistic } from './statistic';
 export { measureTextWidth } from './measure-text';
 export { isBetween, isRealNumber } from './number';
 export * from './data';
+export * from './padding';